+++ app/scripts/components/reduce-select.ts
@@ -8,7 +8,6 @@ type ReduceSelectScope = IScope & {
     keyItems: Record<string, Item>
     hasWordAttrs: boolean
     hasStructAttrs: boolean
-    isOnlyWordSelected: boolean
     onDropdownToggle: (open: boolean) => void
     toggleSelected: (value: string, event: MouseEvent) => void
     toggleWordInsensitive: (event: MouseEvent) => void
@@ -49,12 +48,7 @@ angular.module("korpApp").component("reduceSelect", {
                     ng-class="keyItems['word'].selected ? 'selected':''"
                     class="attribute"
                 >
-                    <input
-                        type="checkbox"
-                        class="reduce-check"
-                        ng-checked="keyItems['word'].selected"
-                        ng-disabled="isOnlyWordSelected"
-                    />
+                    <input type="checkbox" class="reduce-check" ng-checked="keyItems['word'].selected" />
                     <span class="reduce-label">{{keyItems['word'].label | locObj:$root.lang }}</span>
                     <span
                         ng-class="keyItems['word'].insensitive ? 'selected':''"
@@ -111,8 +105,6 @@ angular.module("korpApp").component("reduceSelect", {
                 for (const name of $ctrl.insensitive || []) {
                     if (name in scope.keyItems) scope.keyItems[name].insensitive = true
                 }
-
-                updateIsOnlyWordSelected()
             }
 
             /** Report any changes upwards */
@@ -130,11 +122,6 @@ angular.module("korpApp").component("reduceSelect", {
 
                 // Only notify if something changed
                 if (changes.selected || changes.insensitive) $ctrl.onChange(changes)
-            }
-
-            /** Update the local flag to whether only the word attr is selected. */
-            function updateIsOnlyWordSelected() {
-                scope.isOnlyWordSelected = $ctrl.items?.every((item) => item.value == "word" || !item.selected)
             }
 
             /** Fix state inconsistencies */
@@ -161,10 +148,8 @@ angular.module("korpApp").component("reduceSelect", {
                     item.selected = true
                 }
                 // Toggle given value, unless it is "word" and it is the only one selected.
-                else if (value != "word" || !scope.isOnlyWordSelected) {
+                else {
                     item.selected = !item.selected
-                    // Make sure the word option is locked if it's the only one remaining.
-                    if (scope.keyItems["word"].selected) updateIsOnlyWordSelected()
                 }
             }